+++ tests/impl/try_unzip.rs
@@ -6,8 +6,11 @@ fn try_unzip_fail_a() {
     use collect_failable::TryUnzip;
     use std::collections::HashSet;
 
-    let err_a = INVALID_DATA_A.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should be Err").unwrap_a();
-    assert_eq!(err_a.error.item, 1, "Should be colliding value");
+    let err = INVALID_DATA_A.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should be Err");
+    let data = err.into_data();
+    let side = data.side.left().expect("Should be left");
+
+    assert_eq!(side.error.item, 1, "Should be colliding value");
 }
 
 #[test]
@@ -15,8 +18,11 @@ fn try_unzip_fail_b() {
     use collect_failable::TryUnzip;
     use std::collections::HashSet;
 
-    let err_b = INVALID_DATA_B.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should be Err").unwrap_b();
-    assert_eq!(err_b.error.item, 2, "Should be colliding value");
+    let err = INVALID_DATA_B.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should be Err");
+    let data = err.into_data();
+    let side = data.side.right().expect("Should be right");
+
+    assert_eq!(side.error.item, 2, "Should be colliding value");
 }
 
 #[test]
@@ -26,11 +32,14 @@ fn try_unzip_recover_partial_b_on_a_failure() {
 
     // When FromA fails, the incomplete FromB should be recoverable
     let data = vec![(1, 10), (2, 20), (1, 30), (3, 40)]; // Collision on first element (A)
-    let err_a = data.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should fail").unwrap_a();
-    assert_eq!(err_a.error.item, 1);
-    assert_eq!(err_a.failed, HashSet::from([1, 2])); // Failed collection A
-    assert_eq!(err_a.successful, HashSet::from([10, 20])); // Successful collection B
-    assert_eq!(err_a.remaining.len(), 1);
+    let err = data.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should fail");
+    let err_data = err.into_data();
+    let side = err_data.side.left().expect("Should be left");
+
+    assert_eq!(side.error.item, 1);
+    assert_eq!(side.failed, HashSet::from([1, 2])); // Failed collection A
+    assert_eq!(side.successful, HashSet::from([10, 20])); // Successful collection B
+    assert_eq!(err_data.remaining.len(), 1);
 }
 
 #[test]
@@ -40,10 +49,12 @@ fn try_unzip_recover_partial_a_on_b_failure() {
 
     // When FromB fails, the incomplete FromA should be recoverable
     let data = vec![(10, 1), (20, 2), (30, 2), (40, 3)]; // Collision on second element (B)
-    let err_b = data.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should fail").unwrap_b();
+    let err = data.into_iter().try_unzip::<_, _, HashSet<_>, HashSet<_>>().expect_err("Should fail");
+    let err_data = err.into_data();
+    let side = err_data.side.right().expect("Should be right");
 
-    assert_eq!(err_b.error.item, 2);
-    assert_eq!(err_b.failed, HashSet::from([1, 2])); // Failed collection B
-    assert_eq!(err_b.successful, HashSet::from([10, 20, 30])); // Successful collection A
-    assert_eq!(err_b.remaining.len(), 1);
+    assert_eq!(side.error.item, 2);
+    assert_eq!(side.failed, HashSet::from([1, 2])); // Failed collection B
+    assert_eq!(side.successful, HashSet::from([10, 20, 30])); // Successful collection A
+    assert_eq!(err_data.remaining.len(), 1);
 }